Moved SegoeWpFontResolver
